Created Main component
